muchas cosas hechas

+++ app/src/main/java/com/example/myapplication/Receta.kt
@@ -2,11 +2,11 @@ data class Receta(
     val nombreReceta: String = "",
     val dificultad: String = "",
     val porciones: Int = 0,
-    val tiempoPreparacion: Int = 0,
+    val tiempoPreparacion: Float = 0F,
     val preparacion: String = "",
     val ingredientes: List<String> = emptyList(),
     val imagenUrl: String = "",
     val recetaId: Int = 0,
-    val calorias: Int = 0
+    val calorias: Float = 0F
 
 )
